+++ include/merge_pcl/merge_pcl.h
@@ -8,6 +8,7 @@
 #include <thread>
 #include <deque>
 #include <mutex>
+#include <vector>
 #include <ros/ros.h>
 #include <ros/rate.h>
 #include <pcl_ros/point_cloud.h>
@@ -16,47 +17,45 @@
 #include <exception>
 #include <tf/transform_listener.h>
 #include <pcl_ros/transforms.h>
-#include <pcl/point_types.h>
 #include <pcl/kdtree/kdtree_flann.h>
-#include <message_filters/subscriber.h>
 #include <sensor_msgs/PointCloud2.h>
+
+typedef pcl::PointXYZ PointT;
+typedef pcl::PointCloud<PointT> PointCloudT;
 
 using namespace std;
 
 class PointCloudMerger {
 public:
-    PointCloudMerger();
+    PointCloudMerger(ros::NodeHandle nh);
     ~PointCloudMerger() = default;
 
     // Method to process buffers
     void processBuffers();
 
 private:
-    ros::NodeHandle nh;
-
-    int pcl2_source_num;
+    // Subscribers
+    std::vector<std::string> pc2_topics_in;
     size_t queue_size;
+    std::vector<ros::Subscriber> pc2_subs;
+
     double timeout;
 
-    std::string pcl2_topic_0, pcl2_frame_0;
-    ros::Subscriber pcl2_sub_0;
-    void pcl2Callback0(const sensor_msgs::PointCloud2ConstPtr& pcl2);
-
-    std::string pcl2_topic_1, pcl2_frame_1;
-    ros::Subscriber pcl2_sub_1;
-    void pcl2Callback1(const sensor_msgs::PointCloud2ConstPtr& pcl2);
+    // Transform listener
+    double tf_duration;
+    tf::TransformListener tf_listener;
 
     // Buffers to store point clouds
-    std::deque<pcl::PointCloud<pcl::PointXYZ>::Ptr> buffer_0;
-    std::deque<pcl::PointCloud<pcl::PointXYZ>::Ptr> buffer_1;
+    std::deque<PointCloudT::Ptr> pcT_buffer;
     std::mutex buffer_mutex;
 
-    // Transform listeners
-    tf::TransformListener listener0, listener1;
-    double tf_duration;
+    // Publisher
+    std::string pc2_topic_out;
+    std::string pc2_frame_out;
+    ros::Publisher pc2_pub;
 
-    std::string pcl2_topic_out, pcl2_frame_out;
-    ros::Publisher pcl2_pub;
+    // Unified callback for point cloud subscribers
+    void pc2Callback(const sensor_msgs::PointCloud2ConstPtr& pc2, const size_t index);
 };
 
 #endif //MERGE_PCL_H